chore: upgrade rudder-iac to v0.22.0 (#299)
Move the provider off the pinned pre-release rudder-iac commit (#625) to the
released v0.22.0, which is required to access the destination-side transformation
endpoints used by the destination↔transformation link feature.

The upgrade carries a few mechanical, behaviour-preserving adaptations to
rudder-iac API changes between the old pin and v0.22.0:

- retl: CreateRETLConnectionRequest.SyncBehaviour became *SyncBehaviour — take
  the address (value, including "", is still sent). Test literals wrapped via a
  ptr() helper.
- generatetf: Destination.Version widened to int64 — cast at the call sites.
- destination: populate Destination.Version via int64(cm.Version); existing
  version tests assert int64.
- testutil/acc: compare against int64(wantVersion).

No behavioural change; build + unit tests green.


🔒 Scanned for secrets using gitleaks 8.28.0

+++ rudderstack/resource_destination_test.go
@@ -237,7 +237,7 @@ func TestPopulateDestinationFromState_SetsVersionFromConfigMeta(t *testing.T) {
 	require.NoError(t, populateDestinationFromState(cm, destination, d))
 
 	assert.Equal(t, "TEST", destination.Type)
-	assert.Equal(t, 2, destination.Version)
+	assert.Equal(t, int64(2), destination.Version)
 }
 
 func TestPopulateDestinationFromState_NoVersionSetsZero(t *testing.T) {
@@ -260,5 +260,5 @@ func TestPopulateDestinationFromState_NoVersionSetsZero(t *testing.T) {
 	destination := &client.Destination{}
 	require.NoError(t, populateDestinationFromState(cm, destination, d))
 
-	assert.Equal(t, 0, destination.Version)
+	assert.Equal(t, int64(0), destination.Version)
 }
